refactor: split lexical editor helpers
