+++ src/primitives/lexicalSceneDocumentEditor.js
@@ -1,5 +1,4 @@
 import {
-  $createRangeSelection,
   $createParagraphNode,
   $createTextNode,
   $getNearestNodeFromDOMNode,
@@ -7,10 +6,8 @@ import {
   $getRoot,
   $getSelection,
   $isElementNode,
-  $isLineBreakNode,
   $isRangeSelection,
   $isTextNode,
-  $setSelection,
   COMMAND_PRIORITY_HIGH,
   KEY_ENTER_COMMAND,
   KEY_ESCAPE_COMMAND,
@@ -70,6 +67,25 @@ import {
   unpatchDocumentGetSelection,
   unpatchWindowGetSelection,
 } from "./lexicalRichTextShared.js";
+import {
+  applySelectionToLineNode,
+  clearSelectionTextFormatting,
+  createCollapsedRangeAtPosition,
+  getLexicalOffsetBeforeNode,
+  getLexicalTextLength,
+  setSelectionFromRange,
+} from "./lexicalSceneDocumentSelection.js";
+import {
+  getAdjacentReferenceNodeForCollapsedSelection,
+  getReferenceElementFromContextEvent,
+  getReferenceRichTextStateFromNode,
+  getReferenceSelectionInfo,
+  getReferenceSnapshotFromMentionNode,
+  getTextStyleSegmentElementFromContextEvent,
+  isCollapsedReferenceCaretMovingIntoNode,
+  placeCaretAroundReferenceNode,
+  selectReferenceNodeAsElement,
+} from "./lexicalSceneDocumentReferences.js";
 
 const DEFAULT_PLACEHOLDER = "";
 const LEFT_GUTTER_WIDTH_WITH_NUMBERS = 60;
@@ -577,292 +593,6 @@ const createShell = () => {
   `;
 
   return template.content.cloneNode(true);
-};
-
-const walkTextUnits = (node, visitor) => {
-  if (
-    node.nodeType === Node.ELEMENT_NODE &&
-    node.matches?.('[data-rvn-mention="true"][data-rvn-reference-key]')
-  ) {
-    return visitor({
-      length: node.textContent?.length ?? 0,
-      setRangeAtOffset: (range, offset) => {
-        const parent = node.parentNode;
-        const childIndex = Array.from(parent.childNodes).indexOf(node);
-        const pointOffset = childIndex + (offset > 0 ? 1 : 0);
-        range.setStart(parent, pointOffset);
-        range.setEnd(parent, pointOffset);
-      },
-    });
-  }
-
-  if (node.nodeType === Node.TEXT_NODE) {
-    return visitor({
-      length: node.textContent.length,
-      setRangeAtOffset: (range, offset) => {
-        range.setStart(node, offset);
-        range.setEnd(node, offset);
-      },
-    });
-  }
-
-  if (node.nodeType === Node.ELEMENT_NODE && node.nodeName === "BR") {
-    return visitor({
-      length: 1,
-      setRangeAtOffset: (range, offset) => {
-        const parent = node.parentNode;
-        const childIndex = Array.from(parent.childNodes).indexOf(node);
-        const pointOffset = childIndex + (offset > 0 ? 1 : 0);
-        range.setStart(parent, pointOffset);
-        range.setEnd(parent, pointOffset);
-      },
-    });
-  }
-
-  for (const child of node.childNodes) {
-    if (walkTextUnits(child, visitor)) {
-      return true;
-    }
-  }
-
-  return false;
-};
-
-const createCollapsedRangeAtPosition = (element, targetPosition) => {
-  const range = document.createRange();
-  const normalizedTargetPosition = Math.max(0, Number(targetPosition) || 0);
-  let currentPosition = 0;
-  let lastUnit;
-  let actualPosition = 0;
-  let foundNode = false;
-
-  walkTextUnits(element, (unit) => {
-    const nodeLength = unit.length;
-    lastUnit = unit;
-
-    if (currentPosition + nodeLength >= normalizedTargetPosition) {
-      const offset = normalizedTargetPosition - currentPosition;
-      unit.setRangeAtOffset(range, offset);
-      actualPosition = normalizedTargetPosition;
-      foundNode = true;
-      return true;
-    }
-
-    currentPosition += nodeLength;
-    return false;
-  });
-
-  if (!foundNode && lastUnit) {
-    lastUnit.setRangeAtOffset(range, lastUnit.length);
-    actualPosition = currentPosition;
-    foundNode = true;
-  }
-
-  if (!foundNode) {
-    range.selectNodeContents(element);
-    range.collapse(true);
-  }
-
-  return {
-    range,
-    actualPosition,
-  };
-};
-
-const clampDomPointOffset = (node, offset) => {
-  const numericOffset = Math.max(0, Number(offset) || 0);
-  if (node.nodeType === Node.TEXT_NODE) {
-    return Math.min(numericOffset, node.textContent?.length ?? 0);
-  }
-
-  return Math.min(numericOffset, node.childNodes?.length ?? 0);
-};
-
-const createSafeRange = (element, range) => {
-  if (
-    !element ||
-    !range ||
-    !range.startContainer?.isConnected ||
-    !range.endContainer?.isConnected ||
-    !element.contains(range.startContainer) ||
-    !element.contains(range.endContainer)
-  ) {
-    return undefined;
-  }
-
-  const nextRange = document.createRange();
-  const startOffset = clampDomPointOffset(
-    range.startContainer,
-    range.startOffset,
-  );
-  const endOffset = clampDomPointOffset(range.endContainer, range.endOffset);
-
-  try {
-    nextRange.setStart(range.startContainer, startOffset);
-    nextRange.setEnd(range.endContainer, endOffset);
-  } catch {
-    return undefined;
-  }
-
-  return nextRange;
-};
-
-const setSelectionFromRange = (element, range) => {
-  const safeRange = createSafeRange(element, range);
-  if (!safeRange) {
-    return;
-  }
-
-  const root = element.getRootNode();
-  let selection = window.getSelection();
-
-  if (root instanceof ShadowRoot && typeof root.getSelection === "function") {
-    selection = root.getSelection() || selection;
-  }
-
-  if (!selection) {
-    return;
-  }
-
-  selection.removeAllRanges();
-  selection.addRange(safeRange);
-};
-
-const getLexicalTextLength = (node) => {
-  if ($isLineBreakNode(node)) {
-    return 1;
-  }
-
-  if ($isTextNode(node)) {
-    return node.getTextContentSize?.() ?? node.getTextContent().length;
-  }
-
-  if ($isElementNode(node)) {
-    return node.getChildren().reduce((total, childNode) => {
-      return total + getLexicalTextLength(childNode);
-    }, 0);
-  }
-
-  return 0;
-};
-
-const getLexicalOffsetBeforeNode = (node, targetKey) => {
-  if (node.getKey() === targetKey) {
-    return { found: true, offset: 0 };
-  }
-
-  if (!$isElementNode(node)) {
-    return { found: false, offset: getLexicalTextLength(node) };
-  }
-
-  let offset = 0;
-
-  for (const childNode of node.getChildren()) {
-    const result = getLexicalOffsetBeforeNode(childNode, targetKey);
-    if (result.found) {
-      return {
-        found: true,
-        offset: offset + result.offset,
-      };
-    }
-
-    offset += result.offset;
-  }
-
-  return { found: false, offset };
-};
-
-const resolvePointAtOffset = (node, offset) => {
-  const targetOffset = Math.max(0, Number(offset) || 0);
-
-  if ($isLineBreakNode(node)) {
-    const parent = node.getParent();
-    if (!parent) {
-      return undefined;
-    }
-
-    return {
-      key: parent.getKey(),
-      offset: node.getIndexWithinParent() + (targetOffset > 0 ? 1 : 0),
-      type: "element",
-    };
-  }
-
-  if ($isTextNode(node)) {
-    const textLength = getLexicalTextLength(node);
-    return {
-      key: node.getKey(),
-      offset: Math.min(targetOffset, textLength),
-      type: "text",
-    };
-  }
-
-  if (!$isElementNode(node)) {
-    return undefined;
-  }
-
-  const children = node.getChildren();
-  let remainingOffset = targetOffset;
-
-  for (const childNode of children) {
-    const childLength = getLexicalTextLength(childNode);
-    if ($isLineBreakNode(childNode) && remainingOffset <= childLength) {
-      return {
-        key: node.getKey(),
-        offset:
-          childNode.getIndexWithinParent() + (remainingOffset > 0 ? 1 : 0),
-        type: "element",
-      };
-    }
-
-    if (remainingOffset <= childLength) {
-      const point = resolvePointAtOffset(childNode, remainingOffset);
-      if (point) {
-        return point;
-      }
-    }
-    remainingOffset -= childLength;
-  }
-
-  const lastChild = children.at(-1);
-  if (lastChild) {
-    return resolvePointAtOffset(lastChild, getLexicalTextLength(lastChild));
-  }
-
-  return {
-    key: node.getKey(),
-    offset: 0,
-    type: "element",
-  };
-};
-
-const applySelectionToLineNode = (lineNode, selectionSnapshot = {}) => {
-  const startPoint = resolvePointAtOffset(lineNode, selectionSnapshot.start);
-  const endPoint = resolvePointAtOffset(
-    lineNode,
-    selectionSnapshot.end ?? selectionSnapshot.start,
-  );
-
-  if (!startPoint || !endPoint) {
-    lineNode.selectStart();
-    return;
-  }
-
-  const selection = $createRangeSelection();
-  selection.anchor.set(startPoint.key, startPoint.offset, startPoint.type);
-  selection.focus.set(endPoint.key, endPoint.offset, endPoint.type);
-  $setSelection(selection);
-};
-
-const clearSelectionTextFormatting = (selection) => {
-  selection.format = 0;
-  selection.setStyle("");
-
-  const anchorNode = selection.anchor.getNode();
-  if ($isTextNode(anchorNode)) {
-    anchorNode.setFormat(0);
-    anchorNode.setStyle("");
-  }
 };
 
 export const LEXICAL_SCENE_DOCUMENT_EDITOR_TAG_NAME =
@@ -1786,33 +1516,15 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
   }
 
   getTextStyleSegmentSnapshotFromContextEvent(event) {
-    const element = this.getTextStyleSegmentElementFromContextEvent(event);
+    const element = getTextStyleSegmentElementFromContextEvent(
+      event,
+      this.refs.editor,
+    );
     if (!element) {
       return undefined;
     }
 
     return this.getTextStyleSegmentSnapshotFromElement(element);
-  }
-
-  getTextStyleSegmentElementFromContextEvent(event) {
-    const selector =
-      '[style*="--rvn-text-style-id"], [style*="--rvn-furigana-text"]';
-    const path = event.composedPath?.() ?? [];
-
-    for (const target of path) {
-      const element =
-        target?.nodeType === Node.TEXT_NODE ? target.parentElement : target;
-      const segmentElement = element?.closest?.(selector);
-      if (segmentElement && this.refs.editor?.contains(segmentElement)) {
-        return segmentElement;
-      }
-
-      if (target === this.refs.editor) {
-        break;
-      }
-    }
-
-    return undefined;
   }
 
   getTextStyleSegmentSnapshotFromElement(element) {
@@ -1858,28 +1570,11 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
     });
   }
 
-  getReferenceElementFromContextEvent(event) {
-    const selector = '[data-rvn-mention="true"][data-rvn-reference-key]';
-    const path = event.composedPath?.() ?? [];
-
-    for (const target of path) {
-      const element =
-        target?.nodeType === Node.TEXT_NODE ? target.parentElement : target;
-      const referenceElement = element?.closest?.(selector);
-      if (referenceElement && this.refs.editor?.contains(referenceElement)) {
-        return referenceElement;
-      }
-
-      if (target === this.refs.editor) {
-        break;
-      }
-    }
-
-    return undefined;
-  }
-
   getReferenceSnapshotFromContextEvent(event) {
-    const element = this.getReferenceElementFromContextEvent(event);
+    const element = getReferenceElementFromContextEvent(
+      event,
+      this.refs.editor,
+    );
     if (!element) {
       return undefined;
     }
@@ -1894,17 +1589,8 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
         return undefined;
       }
 
-      return this.getReferenceSnapshotFromNode(node);
+      return getReferenceSnapshotFromMentionNode(node);
     });
-  }
-
-  getReferenceSnapshotFromNode(node) {
-    const reference = node.getReferenceData();
-    return {
-      nodeKey: node.getKey(),
-      resourceId: reference.resourceId,
-      label: reference.label,
-    };
   }
 
   getReferenceRichTextState(nodeKey) {
@@ -1916,153 +1602,13 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
     }
 
     return this.editor.getEditorState().read(() => {
-      const node = $getNodeByKey(nodeKey);
-      if (!$isMentionNode(node)) {
-        return {
-          textStyleId: undefined,
-          furigana: undefined,
-        };
-      }
-
-      return {
-        textStyleId: getTextStyleIdFromNode(node),
-        furigana: getFuriganaFromNode(node),
-      };
+      return getReferenceRichTextStateFromNode($getNodeByKey(nodeKey));
     });
-  }
-
-  getReferenceSelectionInfo(selection) {
-    if (!$isRangeSelection(selection)) {
-      return undefined;
-    }
-
-    const points = selection.getStartEndPoints();
-    if (!points) {
-      return undefined;
-    }
-
-    const [startPoint, endPoint] = points;
-    if (
-      startPoint.type === "element" &&
-      endPoint.type === "element" &&
-      startPoint.key === endPoint.key &&
-      endPoint.offset === startPoint.offset + 1
-    ) {
-      const parentNode = startPoint.getNode();
-      const childNode = $isElementNode(parentNode)
-        ? parentNode.getChildAtIndex(startPoint.offset)
-        : undefined;
-      if ($isMentionNode(childNode)) {
-        return {
-          node: childNode,
-          isCollapsed: false,
-          isWhole: true,
-        };
-      }
-    }
-
-    const startNode = startPoint.getNode();
-    const endNode = endPoint.getNode();
-    if (!$isMentionNode(startNode) || startNode !== endNode) {
-      return undefined;
-    }
-
-    const textLength = getLexicalTextLength(startNode);
-    const isCollapsed = selection.isCollapsed();
-    return {
-      node: startNode,
-      isCollapsed,
-      isWhole:
-        !isCollapsed &&
-        startPoint.offset === 0 &&
-        endPoint.offset === textLength,
-    };
   }
 
   selectReferenceNodeAsElement(node) {
     this.selectedReferenceNodeKey = node.getKey();
-
-    const parentNode = node.getParent();
-    if (!$isElementNode(parentNode)) {
-      node.select(0, getLexicalTextLength(node));
-      return;
-    }
-
-    const index = node.getIndexWithinParent();
-    parentNode.select(index, index + 1);
-  }
-
-  placeCaretAroundReferenceNode(node, direction) {
-    const parentNode = node.getParent();
-    if (!$isElementNode(parentNode)) {
-      if (direction < 0) {
-        node.selectStart();
-        return;
-      }
-
-      node.selectEnd();
-      return;
-    }
-
-    const referenceIndex = node.getIndexWithinParent();
-    const caretIndex = referenceIndex + (direction > 0 ? 1 : 0);
-    parentNode.select(caretIndex, caretIndex);
-  }
-
-  isCollapsedReferenceCaretMovingIntoNode(selection, node, direction) {
-    if (!$isRangeSelection(selection) || !selection.isCollapsed()) {
-      return false;
-    }
-
-    const point = selection.anchor;
-    if (point.getNode() !== node) {
-      return false;
-    }
-
-    const textLength = getLexicalTextLength(node);
-    if (point.offset > 0 && point.offset < textLength) {
-      return true;
-    }
-
-    return (
-      (direction > 0 && point.offset <= 0) ||
-      (direction < 0 && point.offset >= textLength)
-    );
-  }
-
-  getAdjacentReferenceNodeForCollapsedSelection(selection, direction) {
-    if (!$isRangeSelection(selection) || !selection.isCollapsed()) {
-      return undefined;
-    }
-
-    const point = selection.anchor;
-    const node = point.getNode();
-    let candidate;
-
-    if ($isMentionNode(node)) {
-      candidate = this.isCollapsedReferenceCaretMovingIntoNode(
-        selection,
-        node,
-        direction,
-      )
-        ? node
-        : undefined;
-    } else if (point.type === "text" && $isTextNode(node)) {
-      const textLength = getLexicalTextLength(node);
-      if (direction < 0 && point.offset === 0) {
-        candidate = node.getPreviousSibling();
-      } else if (direction > 0 && point.offset >= textLength) {
-        candidate = node.getNextSibling();
-      }
-    } else if (point.type === "element" && $isElementNode(node)) {
-      const childIndex = direction < 0 ? point.offset - 1 : point.offset;
-      candidate =
-        childIndex >= 0 && childIndex < node.getChildrenSize()
-          ? node.getChildAtIndex(childIndex)
-          : undefined;
-    }
-
-    return $isMentionNode(candidate) ? candidate : undefined;
+    selectReferenceNodeAsElement(node);
   }
 
   handleReferenceArrowNavigation(event) {
@@ -2090,29 +1636,23 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
         const selection = $getSelection();
         if (!$isRangeSelection(selection)) {
           if ($isMentionNode(selectedReferenceNode)) {
-            this.placeCaretAroundReferenceNode(
-              selectedReferenceNode,
-              direction,
-            );
+            placeCaretAroundReferenceNode(selectedReferenceNode, direction);
             nextSelectedReferenceNodeKey = undefined;
             didHandle = true;
           }
           return;
         }
 
-        const referenceSelection = this.getReferenceSelectionInfo(selection);
+        const referenceSelection = getReferenceSelectionInfo(selection);
         if (referenceSelection?.node) {
           const isSelectedReference =
             this.selectedReferenceNodeKey === referenceSelection.node.getKey();
           if (referenceSelection.isWhole || isSelectedReference) {
-            this.placeCaretAroundReferenceNode(
-              referenceSelection.node,
-              direction,
-            );
+            placeCaretAroundReferenceNode(referenceSelection.node, direction);
             nextSelectedReferenceNodeKey = undefined;
           } else if (
             referenceSelection.isCollapsed &&
-            !this.isCollapsedReferenceCaretMovingIntoNode(
+            !isCollapsedReferenceCaretMovingIntoNode(
               selection,
               referenceSelection.node,
               direction,
@@ -2129,17 +1669,16 @@ export class LexicalSceneDocumentEditorElement extends HTMLElement {
         }
 
         if ($isMentionNode(selectedReferenceNode)) {
-          this.placeCaretAroundReferenceNode(selectedReferenceNode, direction);
+          placeCaretAroundReferenceNode(selectedReferenceNode, direction);
           nextSelectedReferenceNodeKey = undefined;
           didHandle = true;
           return;
         }
 
-        const adjacentReference =
-          this.getAdjacentReferenceNodeForCollapsedSelection(
-            selection,
-            direction,
-          );
+        const adjacentReference = getAdjacentReferenceNodeForCollapsedSelection(
+          selection,
+          direction,
+        );
         if (!adjacentReference) {
           return;
         }
